Change AccountId datatype to Long

+++ src/main/java/my/projects/myfinance/controller/AccountController.java
@@ -1,7 +1,7 @@
 package my.projects.myfinance.controller;
 
-import my.projects.myfinance.dto.AccountRequestDto;
 import my.projects.myfinance.dto.AccountDto;
+import my.projects.myfinance.dto.AccountRequestDto;
 import my.projects.myfinance.service.AccountService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
@@ -24,13 +24,13 @@ public class AccountController {
 
     @PostMapping
     @ResponseBody
-    public Integer addAccount(@RequestBody AccountRequestDto request){
+    public Long addAccount(@RequestBody AccountRequestDto request) {
         return accountService.addAccount(request);
     }
 
     @PutMapping
     @ResponseBody
-    public Integer editAccount(@RequestBody AccountRequestDto request){
+    public Long editAccount(@RequestBody AccountRequestDto request) {
         return accountService.editAccount(request);
     }
 }
